intrusive double linked list

+++ include/lob/types.hpp
@@ -10,7 +10,10 @@ enum class Side { BUY, SELL };
 using OrderId = uint64_t;
 using Quantity = uint64_t;
 using Timestamp = uint64_t;
-using Price = double;
+
+// Use integer price in cents/ticks for faster comparison and hashing
+// This avoids floating-point comparison issues and enables better hash performance
+using Price = int64_t;
 
 struct Fill {
     OrderId buy_order_id;
